L'Oréal Data Collection Passion Project
I developed this data preprocessing plan for L'Oréal to collect structured and unstructured data for influencer marketing purposes. With 4 portfolios (Luxe, Consumer Products, Dermatological Beauty, and Professional Products), L'Oréal needs a set of varying strategies and constraints in the selection of their influencers; this plan needs to be revisited and updated based on market needs to avoid falling into the innovator's dilemma. Happy reading!